==> scripts/1_deal_with_nas.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('/Users/jiaming/Desktop/Lab2/datas/framingham.csv')

# deal with NAs 
df.isnull().sum() # check which variable has NAs

# fill NAs in the columns with column mean
df.education = df.education.fillna(1)
df.cigsPerDay =df.cigsPerDay.fillna(df.cigsPerDay.mean())
df.BPMeds =df.BPMeds.fillna(0)
df.totChol =df.totChol.fillna(df.totChol.mean())
df.BMI =df.BMI.fillna(df.BMI.mean())
df.heartRate =df.heartRate.fillna(df.heartRate.mean())
df.glucose  =df.glucose.fillna(df.heartRate.mean())

# recheck whether there are still NAs
df.isnull().sum()

df.to_csv('/Users/jiaming/Desktop/Lab2/datas/framingham_without_nas.csv', index=False)
