==> scripts/2_one-hot_encoding.ipynb <==
import pandas as pd
import numpy as np

data = pd.read_csv('/Users/jiaming/Desktop/Lab2/datas/framingham_without_nas.csv')

data

cat = data.loc[:, ['male', 'education','currentSmoker', 'BPMeds', 'prevalentStroke','prevalentHyp', 'diabetes', 'TenYearCHD']]
cat.head()

# convert them to one hot encoding. 
cat_str = cat.astype(str) # nominal to string
cat_oh = pd.get_dummies(cat_str)
cat_oh = cat_oh.drop('TenYearCHD_0', axis = 1) # drop unnecessry column 
cat_oh.head()

data_new = data.drop(['male', 'education','currentSmoker', 'BPMeds', 'prevalentStroke','prevalentHyp', 'diabetes', 'TenYearCHD'], axis =1)
data_new.head()

data_cb = pd.concat([data_new, cat_oh], axis = 1) #Combine categorical features to dataset
data_cb.head()

# Lets rename target feature
data_cb.rename({'TenYearCHD_1':'target'}, axis=1, inplace=True)
data_cb.head()

cat_oh.shape, data_new.shape, data_cb.shape

data_cb.to_csv('/Users/jiaming/Desktop/Lab2/datas/data_cb.csv', index=False)
