==> scripts/3_pos_neg_splitting.ipynb <==
import pandas as pd
import numpy as np

data = pd.read_csv('/Users/jiaming/Desktop/Lab2/datas/data_cb.csv')

data

# analysize whether the model is severly imbalanced
target = data['target']
target.value_counts() # the model is severly imbalanced

neg = data[data.iloc[:, -1] == 0]
neg = neg.iloc[:, :-1]
neg = neg.sample(n=644, random_state=42)

pos = data[data.iloc[:, -1] == 1]
pos = pos.iloc[:, :-1]

neg.to_csv('/Users/jiaming/Desktop/Lab2/datas/neg.csv', index=False)
pos.to_csv('/Users/jiaming/Desktop/Lab2/datas/pos.csv', index=False)
