==> scripts/4_train_val_test.ipynb <==
import pandas as pd
import numpy as np

neg = pd.read_csv('/Users/jiaming/Desktop/Lab2/datas/neg.csv')
pos = pd.read_csv('/Users/jiaming/Desktop/Lab2/datas/pos.csv')

import torch
from torch.utils.data import TensorDataset, DataLoader, random_split

datas = np.concatenate((pos, neg), axis = 0)
datas = np.expand_dims(datas, axis = -2)
labels = np.concatenate(([1] * pos.shape[0], [0] * neg.shape[0]), axis = 0)
labels = np.expand_dims(labels, -1)
datas_tensor = torch.tensor(datas, dtype=torch.float32)
labels_tensor = torch.tensor(labels, dtype=torch.float32)
dataset = TensorDataset(datas_tensor, labels_tensor)
train_size = int(0.8 * len(dataset))
train_size_1 = int(0.8 * train_size)
val_size = train_size - train_size_1
train_size = train_size_1
test_size = len(dataset) - train_size - val_size
train_dataset, val_dataset, test_dataset = random_split(dataset, [train_size, val_size, test_size])

# save datas
torch.save({
    'train_dataset': train_dataset,  
    'val_dataset': val_dataset,
    'test_dataset': test_dataset
}, '/Users/jiaming/Desktop/Lab2/datas/saved_datasets.pth')
